`/game setcode` -> `/setgamecode`
